+++ R/docx_add.R
@@ -178,7 +178,7 @@ body_add_gg <- function(x, value, width = 6, height = 5, res = 300, style = "Nor
 
 #' @export
 #' @title Add a list of blocks into a 'Word' document
-#' @description add a list of blocks produced by \code{block_list} into
+#' @description add a list of blocks produced by [block_list()] into
 #' into an rdocx object.
 #' @inheritParams body_add_break
 #' @param blocks set of blocks to be used as footnote content returned by
@@ -254,7 +254,7 @@ body_add_par <- function(x, value, style = NULL, pos = "after") {
 
 #' @export
 #' @title Add fpar in a 'Word' document
-#' @description Add an \code{fpar} (a formatted paragraph)
+#' @description Add an [fpar()] (a formatted paragraph)
 #' into an rdocx object.
 #' @param x a docx device
 #' @param value a character
@@ -292,7 +292,7 @@ body_add_par <- function(x, value, style = NULL, pos = "after") {
 #' doc <- body_add_fpar(doc, img_in_par)
 #' print(doc, target = tempfile(fileext = ".docx"))
 #'
-#' @seealso \code{\link{fpar}}
+#' @seealso [fpar()]
 #' @family functions for adding content
 body_add_fpar <- function(x, value, style = NULL, pos = "after") {
   if (!is.null(style)) {
@@ -818,15 +818,32 @@ body_add.block_caption <- function(x, value, ...) {
 
 
 #' @export
-#' @describeIn body_add add a [block_list] object.
+#' @describeIn body_add add a [block_list] object. Use this function
+#' to add a list of block elements (e.g. paragraphs, images, tables)
+#' into a Word document in a more efficient way than with usual
+#' `body_add_*` functions. This function will add several elements
+#' in a faster way because the cursor is not calculated for
+#' each iteraction over the elements, as a consequence the
+#' function only append elements at the end of the document
+#' and does not allow to insert elements at a specific position.
 body_add.block_list <- function(x, value, ...) {
-  if (length(value) > 0) {
-    for (i in seq_along(value)) {
-      x <- body_add(x, value = value[[i]])
+
+  x <- cursor_end(x)
+  cursor_elt <- docx_current_block_xml(x)
+  for (i in rev(seq_along(value))) {
+    str <- to_wml(value[[i]], add_ns = TRUE)
+    xml_elt <- as_xml_document(str)
+    if (is.null(cursor_elt)) {
+      xml_add_child(
+        xml_find_first(x$doc_obj$get(), "/w:document/w:body"),
+        xml_elt,
+        .where = 0
+      )
+    } else {
+      xml_add_sibling(cursor_elt, xml_elt, .where = "after")
     }
   }
-
-  x
+  cursor_end(x)
 }
 
 #' @export