+++ tests/testthat/test-layout-helper.R
@@ -1,3 +1,5 @@
+# get_layout ----
+
 test_that("get_layout works as expected", {
   opts <- options(cli.num_colors = 1) # suppress colors for error message check
   on.exit(options(opts))
@@ -19,6 +21,7 @@ test_that("get_layout works as expected", {
 
   expect_no_error(la <- get_layout(x, 1))
   expect_equal(la$index, 1)
+  expect_error(get_layout(x, 1, layout_by_id = FALSE), regexp = "`layout` must be <character>")
 
   # same layout in several masters
   file <- test_path("docs_dir", "test-three-identical-masters.pptx")
@@ -26,7 +29,7 @@ test_that("get_layout works as expected", {
 
   expect_error(get_layout(x, "xxx", NULL), 'Layout "xxx" does not exist')
   expect_error(get_layout(x, "xxx", "Master_1"), 'Layout "xxx" does not exist in master "Master_1"')
-  expect_error(get_layout(x, "Title Slide"), "Layout exists in more than one master")
+  expect_error(get_layout(x, "Title Slide"), "Layout \"Title Slide\" exists in more than one master")
   expect_error(get_layout(x, "Title Slide", "xxx"), 'master "xxx" does not exist')
 
   expect_no_error(la <- get_layout(x, "Title Slide", "Master_1"))
@@ -34,6 +37,16 @@ test_that("get_layout works as expected", {
 
   expect_no_error(la <- get_layout(x, 1))
   expect_equal(la$index, 1)
+})
+
+
+test_that("is_layout_info", {
+  x <- read_pptx()
+  la <- get_layout(x, "Title Slide")
+  expect_true(is_layout_info(la))
+  expect_false(is_layout_info(NULL))
+  expect_false(is_layout_info(1))
+  expect_false(is_layout_info(list()))
 })
 
 
@@ -73,7 +86,7 @@ test_that("incorrect inputs are detected", {
   # layout not unique
   file <- test_path("docs_dir", "test-three-identical-masters.pptx")
   x <- read_pptx(file)
-  expect_error(get_layout(x, "Title Slide"), "Layout exists in more than one master")
+  expect_error(get_layout(x, "Title Slide"), "Layout \"Title Slide\" exists in more than one master")
 })
 
 
@@ -124,4 +137,30 @@ test_that("get layout from slide", {
   la_reference <- get_layout(x, layout)
   expect_identical(la_current, la_reference)
   expect_identical(la_slide, la_reference)
+})
+
+
+# layout_default ----
+
+test_that("layout_default", {
+  opts <- options(cli.num_colors = 1) # suppress colors for error message check
+  on.exit(options(opts))
+
+  x <- read_pptx()
+  expect_false(has_layout_default(x))
+  la_default <- list(layout = NA, master = NA)
+  expect_identical(x$layout_default, la_default)
+  expect_identical(get_layout_default(x), la_default)
+  expect_identical(layout_default(x, as_list = TRUE), la_default)
+
+  layout <- "Title Slide"
+  x <- layout_default(x, layout)
+  expect_true(has_layout_default(x))
+  expect_identical(x$layout_default, list(layout = layout, master = "Office Theme"))
+
+  x <- layout_default(x, NA)
+  expect_false(has_layout_default(x))
+  expect_identical(get_layout_default(x), list(layout = NA, master = NA))
+
+  expect_error(layout_default(x, "xxx"), 'Layout "xxx" does not exist')
 })