+++ tests/testthat/test-pptx-add.R
@@ -1,7 +1,40 @@
 test_that("add wrong arguments", {
-  doc <- read_pptx()
-  expect_error(add_slide(doc, "Title and blah", "Office Theme"), fixed = TRUE)
-  expect_error(add_slide(doc, "Title and Content", "Office Tddheme"), fixed = TRUE)
+  x <- read_pptx()
+  expect_error(add_slide(x, "Title and XXXX", "Office Theme"), fixed = TRUE)
+  expect_error(add_slide(x, "Title and Content", "Office XXXX"), fixed = TRUE)
+})
+
+
+test_that("default layout", {
+  opts <- options(cli.num_colors = 1) # suppress colors for error message check
+  on.exit(options(opts))
+
+  x <- read_pptx()
+  expect_warning(add_slide(x), regexp = "Calling `add_slide()` without specifying a `layout` is deprecated", fixed = TRUE)
+  expect_no_error(expect_warning(add_slide(x)))
+
+  x <- layout_default(x, "Title Slide")
+  expect_no_error(add_slide(x))
+  expect_no_error(add_slide(x, layout = "Two Content"))
+})
+
+
+test_that("master is inferred", {
+  opts <- options(cli.num_colors = 1) # suppress colors for error message check
+  on.exit(options(opts))
+
+  x <- read_pptx()
+  x <- add_slide(x, "Title Slide")
+  x <- add_slide(x, "Title and Content")
+  expect_equal(length(x), 2)
+
+  file <- test_path("docs_dir", "test-layout-unique-and-dupe.pptx")
+  x <- read_pptx(file)
+  x <- add_slide(x, "Title Slide", "Master_1")
+  x <- add_slide(x, "Title Slide") # name is unique so master=NULL works
+  expect_error(add_slide(x, "Title and Content"), regex = "Layout \"Title and Content\" exists in more than one master")
+  x <- add_slide(x, "Title and Content", "Master_1")
+  expect_equal(length(x), 3)
 })
 
 
@@ -29,7 +62,7 @@ test_that("add ggplot into placeholder", {
   require(ggplot2)
   local_edition(3L)
   doc <- read_pptx()
-  doc <- add_slide(doc)
+  doc <- add_slide(doc, "Title and Content")
   gg_plot <- ggplot(data = iris) +
     geom_point(
       mapping = aes(Sepal.Length, Petal.Length),
@@ -58,7 +91,7 @@ test_that("add base plot into placeholder", {
     barplot(1:5, col = 2:6)
   })
   doc <- read_pptx()
-  doc <- add_slide(doc)
+  doc <- add_slide(doc, "Title and Content")
   doc <- ph_with(
     doc, anyplot,
     location = ph_location_fullsize(),
@@ -85,9 +118,9 @@ test_that("add unordered_list into placeholder", {
   )
 
   doc <- read_pptx()
-  doc <- add_slide(doc)
+  doc <- add_slide(doc, "Title and Content")
   doc <- ph_with(doc, ul1, location = ph_location_type())
-  doc <- add_slide(doc)
+  doc <- add_slide(doc, "Title and Content")
   doc <- ph_with(doc, ul2, location = ph_location_type())
   expect_snapshot_doc(x = doc, name = "pptx-add-ul", engine = "testthat")
 })
@@ -112,7 +145,7 @@ test_that("add block_list into placeholder", {
   )
 
   doc <- read_pptx()
-  doc <- add_slide(doc)
+  doc <- add_slide(doc, "Title and Content")
   doc <- ph_with(doc,
     value = value, location = ph_location_type(),
     level_list = c(1, 2, 3)
@@ -211,7 +244,7 @@ test_that("hyperlink image", {
   img.file <- file.path(R.home(component = "doc"), "html", "logo.jpg")
 
   x <- read_pptx()
-  x <- add_slide(x)
+  x <- add_slide(x, "Title and Content")
   x <- ph_with(x, value = external_img(img.file), location = ph_location(newlabel = "logo"))
   x <- ph_hyperlink(x = x, ph_label = "logo", href = "https://cran.r-project.org")
   outputfile <- tempfile(fileext = ".pptx")