+++ NEWS.md
@@ -1,3 +1,29 @@
+# officer 0.6.9 (development version)
+
+## Features
+
+- New functions `body_append_start_context()`, `write_elements_to_context()` and 
+  `body_append_stop_context()` to append elements at the end of a Word document
+  in a faster way. This is useful when the document is large and the user wants to
+  only append elements at the end of the document. 
+- Refactored `body_add.block_list()` to speed up the process of adding a block 
+  list to a Word document, optimization of the process is made by minimizing call 
+  to `docx_current_block_xml()`.
+
+## Issues
+
+- `add_slide()`: It now suffices to specify a unique `layout`. The `master` will be inferred automatically.  
+  Also, the default values for the `layout` (`"Title and Content"`) and `master` (`"Office Theme"`) arg were removed. 
+  They only applied to the default pptx template used by `read_pptx()`. To avoid breaking changes in old code, 
+  `add_slide()` now throws a deprecation warning, if no layout is specified (either via the `layout` arg or by setting 
+  a default layout). It then uses the `layout` arg's former default value (`"Title and Content"`). This makes sure, 
+  that the behavior of `add_slide()` does not change and avoids backward compatibility issues (#635)
+- `layout_default()` sets a default layout for `add_slide()` (#635).
+- `print.rpptx` now uses pluralization and marks the default layout (if set by `layout_default()`) with an asterisk (#635)
+- Fixed `docx_summary()` when a text run is associated with a text style and 
+  when paragraph or text styles are not registered in the Word document style 
+  table (thanks to Kaspar Meili).
+
 # officer 0.6.8
 
 ## Issues